+++ nautilus_trader/examples/strategies/ema_cross_stop_entry.py
@@ -33,7 +33,7 @@ from nautilus_trader.model.enums import TrailingOffsetType
 from nautilus_trader.model.enums import TriggerType
 from nautilus_trader.model.events.order import OrderFilled
 from nautilus_trader.model.identifiers import InstrumentId
-from nautilus_trader.model.instruments.base import Instrument
+from nautilus_trader.model.instruments import Instrument
 from nautilus_trader.model.orderbook.book import OrderBook
 from nautilus_trader.model.orders.market_if_touched import MarketIfTouchedOrder
 from nautilus_trader.model.orders.trailing_stop_market import TrailingStopMarketOrder