+++ nautilus_trader/serialization/arrow/schema.py
@@ -44,13 +44,13 @@ from nautilus_trader.model.events.order import OrderUpdated
 from nautilus_trader.model.events.position import PositionChanged
 from nautilus_trader.model.events.position import PositionClosed
 from nautilus_trader.model.events.position import PositionOpened
-from nautilus_trader.model.instruments.betting import BettingInstrument
-from nautilus_trader.model.instruments.crypto_future import CryptoFuture
-from nautilus_trader.model.instruments.crypto_perpetual import CryptoPerpetual
-from nautilus_trader.model.instruments.currency_pair import CurrencyPair
-from nautilus_trader.model.instruments.equity import Equity
-from nautilus_trader.model.instruments.future import Future
-from nautilus_trader.model.instruments.option import Option
+from nautilus_trader.model.instruments import BettingInstrument
+from nautilus_trader.model.instruments import CryptoFuture
+from nautilus_trader.model.instruments import CryptoPerpetual
+from nautilus_trader.model.instruments import CurrencyPair
+from nautilus_trader.model.instruments import Equity
+from nautilus_trader.model.instruments import FuturesContract
+from nautilus_trader.model.instruments import OptionsContract
 from nautilus_trader.model.orderbook.data import OrderBookData
 from nautilus_trader.serialization.arrow.serializer import register_parquet
 
@@ -638,7 +638,7 @@ NAUTILUS_PARQUET_SCHEMA = {
             "ts_init": pa.uint64(),
         },
     ),
-    Future: pa.schema(
+    FuturesContract: pa.schema(
         {
             "id": pa.dictionary(pa.int64(), pa.string()),
             "native_symbol": pa.string(),
@@ -656,7 +656,7 @@ NAUTILUS_PARQUET_SCHEMA = {
             "ts_init": pa.uint64(),
         },
     ),
-    Option: pa.schema(
+    OptionsContract: pa.schema(
         {
             "id": pa.dictionary(pa.int64(), pa.string()),
             "native_symbol": pa.string(),